memes dowloaded, block created

+++ app/src/main/java/com/example/myapplication/models/MemeDto.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 data class MemeDto (
 
     @SerializedName("id")
-    val id : Integer,
+    val id : Long,
 
     @SerializedName("title")
     val title : String,
